Refactored ApiKeyFetcher tests
